+++ Classes/Service/Shopware/CategoryClientInterface.php
@@ -35,7 +35,9 @@ interface CategoryClientInterface extends AbstractShopwareApiClientInterface
 {
 
     const ENDPOINT = 'categories';
+
     const CACHE_TAG = 'showpare_category';
+
     const ENTITY_CLASS_NAME = Category::class;
 
     /**