+++ Classes/Service/Shopware/MediaClientInterface.php
@@ -35,7 +35,8 @@ interface MediaClientInterface extends AbstractShopwareApiClientInterface
 {
 
     const ENDPOINT = 'media';
-    const CACHE_TAG = 'showpare_media';
-    const ENTITY_CLASS_NAME = Media::class;
 
+    const CACHE_TAG = 'showpare_media';
+
+    const ENTITY_CLASS_NAME = Media::class;
 }